requirements.txt
fastapi
uvicorn
fastapi[standard]
google-cloud-firestore==2.17.2
python-dotenv==1.0.0
langchain
langchain-openai
langchain-community
langchain-core
nltk
requests
fuzzywuzzy[speedup]
azure-ai-textanalytics==5.2.0
beautifulsoup4
